+++ src/client.ts
@@ -1,15 +1,5 @@
-import { AWSSignerV4, md5, parseXML } from "../deps.ts";
-import type {
-  CreateBucketOptions,
-  DeleteBucketPolicyOptions,
-  GetBucketPolicyOptions,
-  GetBucketPolicyStatusOptions,
-  ListBucketsResponses,
-  Policy,
-  PolicyStatus,
-  PutBucketPolicyOptions,
-  Statement,
-} from "./types.ts";
+import { AWSSignerV4, parseXML } from "../deps.ts";
+import type { CreateBucketOptions, ListBucketsResponses } from "./types.ts";
 import { S3Error } from "./error.ts";
 import { S3Bucket } from "./bucket.ts";
 import { doRequest, encoder } from "./request.ts";
@@ -124,149 +114,6 @@ export class S3 {
     return this.#parseListBucketsResponseXml(xml);
   }
 
-  async putBucketPolicy(
-    bucket: string,
-    options: PutBucketPolicyOptions,
-  ): Promise<void> {
-    const headers: Params = {};
-    const params: Params = {};
-    const json = JSON.stringify(options.policy);
-    const body = encoder.encode(json);
-
-    headers["Content-MD5"] = md5(json);
-
-    if (typeof options.confirmRemoveSelfBucketAccess !== "undefined") {
-      headers["x-amz-confirm-remove-self-bucket-access"] = options
-        .confirmRemoveSelfBucketAccess.toString();
-    }
-    if (options.expectedBucketOwner) {
-      headers["x-amz-expected-bucket-owner"] = options.expectedBucketOwner;
-    }
-
-    params["policy"] = "true";
-
-    const resp = await doRequest({
-      host: this.#host,
-      signer: this.#signer,
-      path: bucket,
-      method: "PUT",
-      headers,
-      params,
-      body,
-    });
-
-    if (resp.status !== 204) {
-      throw new S3Error(
-        `Failed to create policy for bucket "${bucket}": ${resp.status} ${resp.statusText}`,
-        await resp.text(),
-      );
-    }
-
-    // clean up http body
-    await resp.arrayBuffer();
-  }
-
-  async getBucketPolicy(
-    bucket: string,
-    options?: GetBucketPolicyOptions,
-  ): Promise<Policy> {
-    const headers: Params = {};
-    const params: Params = {};
-
-    if (options?.expectedBucketOwner) {
-      headers["x-amz-expected-bucket-owner"] = options.expectedBucketOwner;
-    }
-
-    params["policy"] = "true";
-
-    const resp = await doRequest({
-      host: this.#host,
-      signer: this.#signer,
-      path: bucket,
-      method: "GET",
-      headers,
-      params,
-    });
-
-    if (resp.status !== 200) {
-      throw new S3Error(
-        `Failed to get policy for bucket "${bucket}": ${resp.status} ${resp.statusText}`,
-        await resp.text(),
-      );
-    }
-
-    const result = JSON.parse(await resp.text());
-
-    return this.#parseGetBucketPolicyResult(result);
-  }
-
-  async deleteBucketPolicy(
-    bucket: string,
-    options?: DeleteBucketPolicyOptions,
-  ): Promise<void> {
-    const headers: Params = {};
-    const params: Params = {};
-
-    if (options?.expectedBucketOwner) {
-      headers["x-amz-expected-bucket-owner"] = options.expectedBucketOwner;
-    }
-
-    params["policy"] = "true";
-
-    const resp = await doRequest({
-      host: this.#host,
-      signer: this.#signer,
-      path: bucket,
-      method: "DELETE",
-      headers,
-      params,
-    });
-
-    if (resp.status !== 204) {
-      throw new S3Error(
-        `Failed to delete policy for bucket "${bucket}": ${resp.status} ${resp.statusText}`,
-        await resp.text(),
-      );
-    }
-
-    // clean up http body
-    await resp.arrayBuffer();
-  }
-
-  async getBucketPolicyStatus(
-    bucket: string,
-    options?: GetBucketPolicyStatusOptions,
-  ): Promise<PolicyStatus> {
-    const headers: Params = {};
-    const params: Params = {};
-
-    if (options?.expectedBucketOwner) {
-      headers["x-amz-expected-bucket-owner"] = options.expectedBucketOwner;
-    }
-
-    params["policyStatus"] = "true";
-
-    const resp = await doRequest({
-      host: this.#host,
-      signer: this.#signer,
-      path: bucket,
-      method: "GET",
-      headers,
-      params,
-    });
-
-    if (resp.status !== 200) {
-      throw new S3Error(
-        `Failed to get policy status for bucket "${bucket}": ${resp.status} ${resp.statusText}`,
-        await resp.text(),
-      );
-    }
-
-    return this.#parseGetBucketPolicyStatusResult(
-      await resp.text(),
-    );
-  }
-
   #parseListBucketsResponseXml(xml: string): ListBucketsResponses {
     const doc: Document = parseXML(xml);
     const root = extractRoot(doc, "ListAllMyBucketsResult");
@@ -287,42 +134,5 @@ export class S3 {
         displayName: extractContent(owner, "DisplayName"),
       },
     };
-  }
-
-  // deno-lint-ignore no-explicit-any
-  #parseGetBucketPolicyResult(result: Record<string, any>): Policy {
-    const policy: Policy = {
-      statement: Array.isArray(result.Statement)
-        ? result.Statement.map(mapKeys) as Array<Statement>
-        : mapKeys(result.Statement) as Statement,
-    };
-
-    if (result.ID) {
-      policy.id = result.ID;
-    }
-
-    if (result.Version) {
-      policy.version = result.Version;
-    }
-
-    return policy;
-
-    // deno-lint-ignore no-explicit-any
-    function mapKeys(obj: Record<string, unknown>): Record<string, any> {
-      const mapped: Record<string, unknown> = {};
-      for (const key in obj) {
-        const mappedKey = key.slice(0, 1).toLowerCase() + key.slice(1);
-        mapped[mappedKey] = obj[key];
-      }
-      return mapped;
-    }
-  }
-
-  #parseGetBucketPolicyStatusResult(xml: string): PolicyStatus {
-    const doc: Document = parseXML(xml);
-    const root = extractRoot(doc, "PolicyStatus");
-    const isPublic =
-      extractContent(root, "IsPublic ")?.toLowerCase() === "true";
-    return { isPublic };
   }
 }
