+++ src/client_test.ts
@@ -3,7 +3,6 @@ import { S3Error } from "./error.ts";
 import { S3Bucket } from "./bucket.ts";
 import { S3 } from "./client.ts";
 import { encoder } from "./request.ts";
-import { Policy } from "./types.ts";
 
 const s3 = new S3({
   accessKeyID: Deno.env.get("AWS_ACCESS_KEY_ID")!,
@@ -11,25 +10,6 @@ const s3 = new S3({
   region: "us-east-1",
   endpointURL: Deno.env.get("S3_ENDPOINT_URL"),
 });
-
-const policy: Policy = {
-  version: "2012-10-17",
-  id: "test",
-  statement: [
-    {
-      effect: "Allow",
-      principal: {
-        AWS: ["111122223333", "444455556666"],
-      },
-      action: [
-        "s3:PutObject",
-      ],
-      resource: [
-        "arn:aws:s3:::*",
-      ],
-    },
-  ],
-};
 
 Deno.test({
   name: "[client] should create a new bucket",
@@ -74,53 +54,5 @@ Deno.test({
       "creationDate is not of type Date",
     );
     assertEquals(owner.displayName, "minio");
-  },
-});
-
-Deno.test({
-  name: "[client] should put a bucket policy",
-  async fn() {
-    await s3.putBucketPolicy("test.bucket", { policy });
-
-    // teardown
-    await s3.deleteBucketPolicy("test.bucket");
-  },
-});
-
-Deno.test({
-  name: "[client] should get a bucket policy",
-  async fn() {
-    await s3.putBucketPolicy("test.bucket", { policy });
-    const resp = await s3.getBucketPolicy("test.bucket");
-    assertEquals(resp, policy);
-
-    // teardown
-    await s3.deleteBucketPolicy("test.bucket");
-  },
-});
-
-Deno.test({
-  name: "[client] should delete a bucket policy",
-  async fn() {
-    await s3.putBucketPolicy("test.bucket", { policy });
-    const resp = await s3.getBucketPolicy("test.bucket");
-    assert(resp);
-
-    // teardown
-    await s3.deleteBucketPolicy("test.bucket");
-
-    await assertThrowsAsync(
-      () => s3.getBucketPolicy("test.bucket"),
-      S3Error,
-      'Failed to get policy for bucket "test.bucket": 404 Not Found',
-    );
-  },
-});
-
-Deno.test({
-  name: "[client] should get the bucket policy status",
-  async fn() {
-    const resp = await s3.getBucketPolicyStatus("test.bucket");
-    assertEquals(resp, { isPublic: false });
   },
 });
